% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_qpcr_results}
\alias{calculate_qpcr_results}
\title{计算qPCR结果}
\usage{
calculate_qpcr_results(sample_df, gene_names)
}
\arguments{
\item{sample_df}{样本数据框}

\item{gene_names}{基因名称向量}
}
\value{
包含结果和摘要的列表
}
\description{
计算qPCR结果
}
\keyword{internal}
