% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{get_valid_gene_name}
\alias{get_valid_gene_name}
\title{获取有效基因名称}
\usage{
get_valid_gene_name(prompt)
}
\arguments{
\item{prompt}{提示信息}
}
\value{
有效的基因名称
}
\description{
获取有效基因名称
}
\keyword{internal}
