% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{get_valid_numeric_vector}
\alias{get_valid_numeric_vector}
\title{获取有效数值向量}
\usage{
get_valid_numeric_vector(prompt, expected_length)
}
\arguments{
\item{prompt}{提示信息}

\item{expected_length}{期望的数值长度}
}
\value{
有效的数值向量
}
\description{
获取有效数值向量
}
\keyword{internal}
